requirements.txt
langchain
langchain-ollama
langchain-core
speech-recognition
pyttsx3
python-dotenv
duckduckgo-search
pillow
pytesseract
mss
pytz
